FIX: prevent panic when line does not contain ':'
It's possible that the parsing of the output did not contain a colon.
This should possibly be handled elsewhere, but as a stogap measure,
wrong input should not be able to panic from within this function.

output which triggered bug was a permission error:

Potential error parsing output of /Users/tim/mlog/go ; couldn't handle
this stray line: go install mlog/util: open
/Users/tim/mlog/go/pkg/darwin_amd64/mlog/util.a: permission denied