Implemented failure-mode settings per Convey-block
This allows tests to optionally specify on a block-level if they
want their failing So()-assertions to block further execution of
the test arm or if they want failing So()-assertions to allow the
test-arm to proceed.

Top-level blocks default to halting execution on a So()-assertion
failure and every block inherits from its parent-block to avoid
repetition.