Minimal fixes to reporting for nesting scopes.
Uses a quick hack to correctly obtain a 'unique' ID for scopes
in the flat map used by the JSON and story reports. A correctly
crafted Convey name could collide with a key in the map, but that
is likely to be self-sabotage.

A more elaborate solution would be to have a generic report-gatherer
tree, and then implement the various output reports as walkers over
the tree structure at the end of every story.