Fixing a previously untriggered error
My previous fix started reporting errors that were previously ignored.  As a result,
this test case (which should have failed before, but didn't) started to fail.  It turns
out that the directory referenced in the test was incorrect but because the errors
were not caught previously, it seemed to pass.  Once my fix was applied, the test
failed (as it should have).  This commit corrects the directory name so the test
now passes again.