Trying to address build issue
The builds are failing with complains about coverage tools not being installed.  I suspect
that the lack of error checking previously either masked an existing error or intentionally
allowed failures and compensated elsewhere with other logic.  To test this, I'm reverting
once of my changes to see if it allows the build to succeed.  My fear is this will continue
to mask an error (or at least fail to clarify the expected behavior).  But even after
reverting this change, other (more important aspects, at least to me) are still addressed
by my changes.