Keep supporting Go versions < 1.9
The go/types package introduced a change in Go 1.9 (not documented in
the relese notes) that is required for golang.org/x/tools/go/packages
to compile. The previous commit's dependency on that library would
break support for Go versions < 1.9. Since those versions can't be
using modules, the old implementation using go/build will continue to
work for them. We separate the impacted functions out into two versions
that use build tags with conditional compilation to use the out of
GOPATH module supporting version in Go >= 1.9 and the old behavior in
Go < 1.9.