upgrade deps, update to go 1.18
go 1.18 is the most recent supported version of Go, so it seems reasonable
to upgrade to that. This upgrade is needed to pull in the latest version
of smartystreets/assertions, which I want in here for this fix of mine:
https://github.com/smartystreets/assertions/commit/a83f75f344d7f8c0681e5c699c5f747cd41cd93f

This also replaces all "interface{}" with "any", since go 1.18 added that.
That was already done in the assertions repo the other day.

Note that this command passes:
go test -timeout=1s -race -cover -short -count=1 ./...

but the integration test file fails. That failure is already in the repo
at HEAD though.