fix: fix a few tests that were broken at HEAD
one of these tests double-counted goroutines, because logging output
now includes "goroutine" twice. This looks for just the first instance.

the other test failed on test instances which used multiple "%" signs
after coverage numbers. It works with this change, and IMO it's simpler
than it was as well.

also update a gitignore to use a correct new path